% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of x and y
}
\description{
Add two numbers
}
\examples{
add_numbers(1, 2)
}
