% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{create_base_theme_elements}
\alias{create_base_theme_elements}
\title{Create base theme elements}
\usage{
create_base_theme_elements(
  base_font = "Open Sans",
  title_font = "Roboto",
  base_size = 11,
  colors = aiddata_colors,
  grid = TRUE
)
}
\arguments{
\item{base_font}{Base font family}

\item{title_font}{Title font family}

\item{base_size}{Base font size}

\item{colors}{Color palette to use}

\item{grid}{Show grid lines}
}
\value{
A list of theme elements
}
\description{
Create base theme elements
}
\keyword{internal}
