% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library where the package is installed}

\item{pkgname}{The package name}
}
\description{
Checks for required fonts and displays message if they're missing.
}
\keyword{internal}
