% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{download_font_variant}
\alias{download_font_variant}
\title{Download a single font variant}
\usage{
download_font_variant(url, dest_path)
}
\arguments{
\item{url}{Font file URL}

\item{dest_path}{Destination path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Download a single font variant
}
\keyword{internal}
