% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{download_font_variants}
\alias{download_font_variants}
\title{Download all variants of a font}
\usage{
download_font_variants(font_info, font_name)
}
\arguments{
\item{font_info}{Font information from Google Fonts}

\item{font_name}{Name of the font}
}
\value{
Character vector of downloaded file paths
}
\description{
Download all variants of a font
}
\keyword{internal}
