% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{fetch_google_font_info}
\alias{fetch_google_font_info}
\title{Download font information from Google Fonts API}
\usage{
fetch_google_font_info(
  font_name,
  api_key = "AIzaSyDOr3jWLtl4IP08yNaddV61_40f0YByPHo"
)
}
\arguments{
\item{font_name}{Name of font to download}

\item{api_key}{Google Fonts API key}
}
\value{
List containing font information or NULL if not found
}
\description{
Download font information from Google Fonts API
}
\keyword{internal}
