% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{format_installation_message}
\alias{format_installation_message}
\title{Format installation results message}
\usage{
format_installation_message(success)
}
\arguments{
\item{success}{Whether installation was successful}
}
\value{
A character string with the appropriate message
}
\description{
Format installation results message
}
\keyword{internal}
