% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_available_font}
\alias{get_available_font}
\title{Find an available font from a list of options}
\usage{
get_available_font(font_family, fallbacks = c("Arial", "Helvetica", "sans"))
}
\arguments{
\item{font_family}{Primary font family to use}

\item{fallbacks}{Vector of fallback font families}
}
\value{
Name of first available font, or "sans" if none found
}
\description{
Find an available font from a list of options
}
