% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_font_directory}
\alias{get_font_directory}
\title{Get system font directory}
\usage{
get_font_directory()
}
\value{
Path to system font directory
}
\description{
Get system font directory
}
\keyword{internal}
