% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_font_family}
\alias{get_font_family}
\title{Get font family with fallbacks}
\usage{
get_font_family(
  font_family,
  fallbacks = c("Arial", "Helvetica", "sans-serif", "sans")
)
}
\arguments{
\item{font_family}{The primary font family to use}

\item{fallbacks}{Character vector of fallback fonts}
}
\value{
A font family name that is available on the system
}
\description{
Get font family with fallbacks
}
