% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_aiddata_fonts}
\alias{install_aiddata_fonts}
\title{Install AidData fonts}
\usage{
install_aiddata_fonts(force_reinstall = FALSE)
}
\arguments{
\item{force_reinstall}{If TRUE, will attempt to reinstall even
if fonts are present}
}
\value{
Logical indicating whether all fonts were successfully
installed
}
\description{
Downloads and installs the fonts required for AidData themes
(Roboto and Open Sans) from Google Fonts. This requires an internet
connection and permission to install fonts on your system.
}
\examples{
\dontrun{
install_aiddata_fonts()
}
}
