% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_font_files}
\alias{install_font_files}
\title{Install font files to system}
\usage{
install_font_files(font_files, font_dir)
}
\arguments{
\item{font_files}{Vector of font file paths}

\item{font_dir}{Target font directory}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Install font files to system
}
\keyword{internal}
