% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_single_font}
\alias{install_single_font}
\title{Install a single font}
\usage{
install_single_font(font_name, font_dir, force_reinstall)
}
\arguments{
\item{font_name}{Name of the font to install}

\item{font_dir}{Directory to install fonts to}

\item{force_reinstall}{Whether to force reinstall}
}
\value{
List containing success status and any error message
}
\description{
Install a single font
}
\keyword{internal}
