% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{aiddata_colors}
\alias{aiddata_colors}
\alias{aiddata_palettes}
\title{AidData color system based on William & Mary brand guidelines}
\format{
An object of class \code{list} of length 18.

An object of class \code{list} of length 6.
}
\usage{
aiddata_colors

aiddata_palettes
}
\description{
AidData color system based on William & Mary brand guidelines
}
\keyword{datasets}
