% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_aiddata}
\alias{scale_color_aiddata}
\alias{scale_fill_aiddata}
\title{Create AidData color scales}
\usage{
scale_color_aiddata(palette = "default", reverse = FALSE, discrete = TRUE, ...)

scale_fill_aiddata(palette = "default", reverse = FALSE, discrete = TRUE, ...)
}
\arguments{
\item{palette}{The name of the palette to use. Options are:
"default", "alt", "sequential_green", "sequential_gold",
"diverging_green_gold", "diverging_vine_sky"}

\item{reverse}{Boolean indicating whether to reverse the palette}

\item{discrete}{Boolean indicating whether the scale is discrete or
continuous}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn()}
}
\value{
A ggplot2 color scale
}
\description{
Create AidData color scales
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_aiddata("default")
}
